backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
celery
redis
beautifulsoup4
requests
scrapy
playwright
openai
pdfminer.six
python-docx
numpy
scikit-learn
python-multipart
